backend/requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
joblib==1.4.2
numpy==2.2.2
pandas==2.2.3
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2
